refactor: закончить рефакторинг, начать проверку пропсов
