+++ src/components/app/app.module.css
@@ -10,7 +10,7 @@
   align-items: center;
 }
 
-.main{
+.main {
   box-sizing: border-box;
   width: 100%;
   height: 100%;
@@ -20,3 +20,10 @@
   flex-direction: row;
 }
 
+.sectionClass {
+  flex-direction: column;
+  display: flex;
+  width: 50%;
+  min-width: 640px;
+  box-sizing: border-box;
+}
