+++ src/components/burger-ingredients/burger-ingredients.module.css
@@ -1,65 +1,26 @@
-
 .ingredients {
   width: 100%;
   display: flex;
   align-items: stretch;
   flex-direction: column;
-
 }
 
-.type{
-
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-
-  align-items: stretch;
-  grid-column-gap: 24px;
-  grid-row-gap: 32px;
+.nonList {
+  list-style: none;
+  margin: 0;
+  padding: 0;
 }
 
-.typePart{
+.typePart {
   width: 100%;
   box-sizing: border-box;
 }
 
-.sectionClass {
-  flex-direction: column;
-  display: flex;
-  width: 50%;
-  min-width: 640px;
-  box-sizing: border-box;
-}
-
-.price{
-  display: flex;
-  justify-content: center;
-
-}
-
-
-.card{
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  box-sizing: border-box;
-  align-items: center;
-  overflow: hidden;
-
-}
-
-.cardName{
-  height: 48px;
-  width: 100%;
-  text-align: center;
-}
-
-.imgCard {
-  width: 240px;
-  height: 120px;
-  object-fit: cover;
-}
-
-.nonLink{
+.nonLink {
   text-decoration: none;
+}
+
+.tab {
+  display: flex;
+  margin-bottom: 40px;
 }
