+++ src/components/card/card-buns.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "../burger-ingredients/burger-ingredients.module.css";
+import styles from "./card.module.css";
 import {Counter, CurrencyIcon} from "@ya.praktikum/react-developer-burger-ui-components";
 
 const CardBuns = ({currentItem, setSelectedIngredients, selectedIngredients, defaultBunId}) => {
@@ -12,6 +12,7 @@ const CardBuns = ({currentItem, setSelectedIngredients, selectedIngredients, def
         });
       }
     }
+
     check();
   }, [])
   const count = (currentItem._id === selectedIngredients.bun._id) ? 1 : 0
@@ -28,7 +29,7 @@ const CardBuns = ({currentItem, setSelectedIngredients, selectedIngredients, def
   }
 
   return (
-    <div className={styles.card} onClick={toggleCount}>
+    <li className={styles.card} onClick={toggleCount}>
       <img className={styles.imgCard} alt={currentItem.name} src={currentItem.image}/>
       <div className={`pt-1 pb-1 ${styles.price}`}>
         <p className="text text_type_digits-default pr-2">{currentItem.price}</p>
@@ -38,7 +39,7 @@ const CardBuns = ({currentItem, setSelectedIngredients, selectedIngredients, def
         <p className="text text_type_main-default">{currentItem.name}</p>
       </div>
       {isNum(count) && <Counter count={count} size="default" extraClass="m-1"/>}
-    </div>
+    </li>
   );
 };
