+++ src/components/constructor-list/constructor-list.module.css
@@ -1,0 +1,29 @@
+.listScroll {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+}
+
+.scroll {
+  overflow-y: auto;
+  scroll-behavior: smooth;
+}
+
+.scroll::-webkit-scrollbar {
+  width: 8px;
+}
+
+.scroll::-webkit-scrollbar-thumb {
+  background: #8585ad;
+  border: none;
+  border-radius: 0;
+}
+
+.elementConstructor {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: 8px;
+  justify-content: flex-start;
+}
