+++ src/components/list-cards/list-cards.module.css
@@ -1,0 +1,10 @@
+.listCards {
+  list-style: none;
+  margin: 0;
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  align-items: stretch;
+  grid-column-gap: 24px;
+  grid-row-gap: 32px;
+  padding: 24px 16px 40px;
+}
